sync

+++ WZAnimatingMedia/WZAnimatingMediaKit/WZAnimationNode.swift
@@ -12,7 +12,7 @@ class WZAnimationNode {
     
     private(set) var keyname: String
     private(set) var inputNode: WZAnimationNode?
-    private(set) var currentFrame = 0
+    private(set) var currentFrame = 0.0
     
     var localPath: WZBezierPath?
     var outputPath: WZBezierPath?
@@ -23,7 +23,7 @@ class WZAnimationNode {
     }
     
     //TODO 现在在rendergroup中把最后一个当做链表的头，更新的时候相当于反向遍历，然后遍历的时候再遍历到最后一个开始向前递归调用。相当于反过来又返回去，待优化成两次都是正向，便于理解
-    func update(frame: Int) {
+    func update(frame: Double) {
         
         inputNode?.update(frame: frame)
